modify migration file

+++ db/migrate/20170413064405_create_users.rb
@@ -2,8 +2,7 @@ class CreateUsers < ActiveRecord::Migration[5.0]
   def change
     create_table :users do |t|
       t.string :name
-      t.integer :score
-
+      t.float :score
       t.timestamps
     end
   end
